Refactor #57 : 코드 리팩토링
