+++ components/my-page/product/edit-product.tsx
@@ -1,8 +1,8 @@
 "use client";
 
 import { ChangeEvent, DragEvent, MouseEvent, useEffect, useState } from "react";
-import Input from "../global/input";
-import { ProductLabel } from "./label";
+import Input from "../../global/input";
+import { ProductLabel } from "../label";
 import {
 	PRODUCT_CATEGORIES,
 	PRODUCT_CATEGORIES_REVERSE,
@@ -19,12 +19,15 @@ import { IProduct } from "@/model/product";
 import { useRouter } from "next/navigation";
 import useAuth from "@/libs/auth";
 import { IResponse } from "@/model/responses";
-import Modal from "../global/modal";
+import Modal from "../../global/modal";
 import { useRecoilState } from "recoil";
 import { showDefaultModalState } from "@/recoil/atoms";
 import { ISizeTable } from "@/constants/type-table";
 import { getSizeTable } from "@/libs/utils/table";
-import SizeTable from "../global/size-table";
+import SizeTable from "../../global/size-table";
+import { FormModal } from "./form-modal";
+import { ProductDetails } from "./product-details";
+import { ImageUploader } from "./image-uploader";
 
 export const base64ToFile = (
 	base64Data: string,
@@ -42,7 +45,7 @@ export const base64ToFile = (
 	return new File([blob], fileName, { type: contentType });
 };
 
-export default function ProductForm(data: any) {
+export default function EditProductForm(data: any) {
 	const [selectedType, setSelectedType] = useState<string | null>(null);
 	const [selectedSize, setSelectedSize] = useState<string | null>(null);
 	const [selectedGender, setSelectedGender] = useState<string | null>(null);
@@ -222,164 +225,45 @@ export default function ProductForm(data: any) {
 
 	return (
 		<>
-			<Modal onClose={handleCloseModal}>{isShowModal && modalContent}</Modal>
-
-			{sizeTable && isShowSize && (
-				<Modal onClose={handleCloseSize}>
-					<SizeTable
-						tableHeader={sizeTable!.header}
-						tableBody={sizeTable!.body}
-					></SizeTable>
-				</Modal>
-			)}
-
+			<FormModal
+				isShowModal={isShowModal}
+				isShowSize={isShowSize}
+				modalContent={modalContent}
+				sizeTable={sizeTable}
+				handleCloseModal={handleCloseModal}
+				handleCloseSize={handleCloseSize}
+			/>
 			<div className="h-fit flex flex-col justify-start items-start px-4 lg:px-36 pt-60 max-w-screen-2xl">
 				<div className="display">상품 정보 수정</div>
 				<form
 					className="w-full flex flex-col mt-16 max-w-screen-xl"
 					onSubmit={handleSubmitUpdate}
 				>
-					<div className="w-full">
-						<div className="label-1 flex flex-col justify-between w-full mb-4 lg:flex-row">
-							<span>상품 이미지(최대 5장)*</span>
-							<span className="text-gray-400 md:mt-0">
-								제일 첫 번째 이미지가 상품의 대표 이미지가 됩니다. 이미지를
-								끌어당겨 순서를 바꿀 수 있습니다.
-							</span>
-						</div>
-						<input
-							type="file"
-							accept="image/*"
-							multiple
-							onChange={handleImageUpload}
-							className="hidden"
-							id="image-upload"
-						/>
-						<label
-							htmlFor="image-upload"
-							className="cursor-pointer"
-							onClick={handleLabelClick}
-						>
-							<div className="grid grid-cols-3 md:grid-cols-4 lg:grid-cols-5 mt-[1rem] w-full gap-3">
-								{Array.from({ length: 5 }).map((_, index) => (
-									<div
-										key={index}
-										className="relative h-[20rem] bg-gray-100 flex justify-center items-center"
-										draggable={!!images[index] || !!imageURLs[index]}
-										onDragStart={(event) => handleDragStart(event, index)}
-										onDrop={(event) => handleDrop(event, index)}
-										onDragOver={handleDragOver}
-									>
-										{images[index] && (
-											<>
-												<img
-													src={URL.createObjectURL(images[index])}
-													alt={`uploaded-${index}`}
-													className="object-cover h-[311px] w-full"
-													style={{ objectFit: "cover" }}
-												/>
-												<button
-													type="button"
-													onClick={(event) => handleRemoveImage(event, index)}
-													className="absolute top-0 right-0 m-2 p-1 size-8 bg-white rounded-full text-gray-500 hover:text-gray-700"
-												>
-													X
-												</button>
-											</>
-										)}
-									</div>
-								))}
-							</div>
-						</label>
-						{errors.productImages && (
-							<span className="text-red-500">{errors.productImages}</span>
-						)}
-					</div>
-					<div className="w-full md:w-[36rem] mt-12 label-1 flex flex-col gap-3">
-						<div>상품 이름 *</div>
-						<Input
-							name="productName"
-							count={40}
-							defaultValue={initialData.title}
-						/>
-						{errors.productName && (
-							<span className="text-red-500">{errors.productName}</span>
-						)}
-					</div>
-					<div className="w-full flex flex-col gap-[0.75rem] label-1 mt-[3.75rem]">
-						<div className="font-bold">상품 종류 (1개 선택) *</div>
-						<ProductLabel
-							list={PRODUCT_CATEGORIES}
-							selectedItems={selectedType ? [selectedType] : []}
-							onSelect={selectType}
-						/>
-					</div>
-					<div className="w-full flex flex-col gap-[0.75rem] label-1">
-						<div className="font-bold flex-col md:flex-row gap-[0.75rem] mt-[3.75rem]">
-							<span>상품 사이즈 *</span>
-							{sizeTable && (
-								<span
-									onClick={handleShowModal}
-									className="label-3 text-gray-400 mt-8 underline cursor-pointer"
-								>
-									사이즈 가이드
-								</span>
-							)}
-						</div>
-						<ProductLabel
-							list={PRODUCT_SIZE} // size 추가 필요
-							selectedItems={selectedSize ? [selectedSize] : []}
-							onSelect={selectSize}
-						/>
-						{errors.productSize && (
-							<span className="text-red-500">{errors.productSize}</span>
-						)}
-					</div>
-					<div className="w-full flex flex-col gap-[0.75rem] label-1">
-						<div className="w-full md:w-[36rem] mt-[3rem] label-1 flex flex-col gap-[12px]">
-							<div>맞춤 키 *</div>
-							<select
-								// 수정 필요
-								defaultValue={selectedHeight || ""}
-								name="height"
-								className="text-gray-800 py-5 px-4 rounded-md focus:outline-none ring-2 focus:ring-4 transition ring-neutral-200 focus:ring-orange-500 border-none"
-							>
-								<option>사이즈를 선택하세요.</option>
-								{Object.entries(PRODUCT_HEIGHT).map(([key, value]) => (
-									<option key={value}>{value}</option>
-								))}
-							</select>
-						</div>
-						{errors.productHeight && (
-							<span className="text-red-500">{errors.productHeight}</span>
-						)}
-					</div>
-					<div className="w-full flex flex-col gap-[0.75rem] label-1">
-						<div className="font-bold flex gap-[0.75rem] mt-[3.75rem]">
-							분류(1개 선택) *
-						</div>
-						<ProductLabel
-							list={PRODUCT_TYPES}
-							selectedItems={selectedGender ? [selectedGender] : []}
-							onSelect={selectGender}
-						/>
-						{errors.productGender && (
-							<span className="text-red-500">{errors.productGender}</span>
-						)}
-					</div>
-					<div className="w-full flex flex-col gap-[0.75rem] label-1">
-						<div className="font-bold flex gap-[0.75rem] mt-[3.75rem]">
-							스타일(최소 1개, 중복 가능) *
-						</div>
-						<ProductLabel
-							list={PRODUCT_STYLES}
-							selectedItems={selectedStyles ? selectedStyles : []}
-							onSelect={toggleStyle}
-						/>
-						{errors.productStyles && (
-							<span className="text-red-500">{errors.productStyles}</span>
-						)}
-					</div>
+					<ImageUploader
+						images={images}
+						imageURLs={imageURLs}
+						onImageUpload={handleImageUpload}
+						onDragStart={handleDragStart}
+						onDrop={handleDrop}
+						onDragOver={handleDragOver}
+						onRemoveImage={handleRemoveImage}
+						onLabelClick={handleLabelClick}
+						errors={errors.productImages}
+					/>
+					<ProductDetails
+						initialData={initialData}
+						selectedType={selectedType}
+						selectedSize={selectedSize}
+						selectedGender={selectedGender}
+						selectedStyles={selectedStyles}
+						errors={errors}
+						selectType={selectType}
+						selectSize={selectSize}
+						selectGender={selectGender}
+						toggleStyle={toggleStyle}
+						handleShowModal={handleShowModal}
+						selectedHeight={selectedHeight}
+					/>
 					<div className="mt-[5rem] w-full h-fit flex justify-center label-1 text-gray-100">
 						<button type="submit" className="bg-primary px-12 py-4 rounded-3xl">
 							상품 수정하기
